market_place/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.6
django-celery==3.1.17
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-js-asset==2.1.0
django-mptt==0.14.0
django-mptt-admin==2.5.1
django-templated-mail==1.1.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gevent==23.9.1
greenlet==3.0.1
idna==3.4
install==1.3.5
kombu==5.3.2
oauthlib==3.2.2
Pillow==10.1.0
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==2.8.0
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
zope.event==5.0
zope.interface==6.1
